+++ indicator-service/src/main/kotlin/com/emgc/indicatorservice/dto/upbit/candle/request/UpbitWeekCandleRequest.kt
@@ -1,4 +1,4 @@
-package com.emgc.indicatorservice.dto.upbit.request
+package com.emgc.indicatorservice.dto.upbit.candle.request
 
 data class UpbitWeekCandleRequest(
     val market: String = "KRW-BTC",
